requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.6
django-cors-headers==3.8.0
django-graphql-jwt==0.3.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
promise==2.3
PyJWT==2.1.0
pytz==2021.1
regex==2021.8.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tomli==1.2.1
